fix(cloud): ensure cloud project id is set on garden class (#5306)
* fix(cloud): ensure cloud project id is set on garden class

This commit fixes a regression introduced with:
https://github.com/garden-io/garden/pull/5291

That commit was just a tiny tweak to a couple of lines of code to
improve error messages but actually ended up breaking the core <> cloud
interaction for the community tier dashboard.

The underlying reason is the fact that this particular code was quite
convoluted and not tested well enough.

So this commit refactors the cloud init logic to make it more readable
and adds a handful of tests to prevent these kind of regressions
from happening again.

* improvement(cloud): better errors when fetching project fails

* fix(core): ensure id and domain are set even if user isn't logged in

A previous commit in the same PR changed this behaviour. This simply
reverts it back to what it was.

* chore: fix tests (TBS)