fix(cloud): correctly send error messages to server
Before this fix, we'd "render" error logs before sending them to
the server which meant that the "error" part was kept but the "msg" part
was dropped.

The user still sees the error but the message can be a bit chopped. This
fixes that by sending the "msg" field as is and adding an "error" field
to the payload as well that cloud handles separately.