fix(k8s): ensure patchResources can patch namespace
When applying K8s resources, we create a ConfigMap with metadata about
the resources for faster lookup.

This metadata includes the resource namespace among other things.

Before this fix, we'd create the ConfigMap before applying the resource
patch which meant that if you patched the actual resource namespace,
Garden wouldn't store that information.

Later when looking up the status of the resource, Garden would use the
data from the ConfigMap and basically look for the resource in the
wrong namespace and return status "not-ready" (when in fact the resource
could be ready, just in a different namespace).