Misc log fixes (#5330)
* fix(log): ensure action log inherits fixLevel

Before this fix, action log entries created from an input log with
'fixLevel' set, would not inherit that level but instead default to the
info level.

This was only really noticeable around some edge in the web dashboard
were logs that should've been verbose were rendered at the info level.

This fixes that and adds corresponding tests.

* improvement(server): log internal command requests at debug level

Before this fix, requests for internal commands like
_get-deploy-statuses weren't logged at all.

Now we log them at the debug level which is useful for, well, debugging
but also makes them visible in the web dashboard (if the appropriate
level is set) where they originate from.

* fix(cloud): correctly send error messages to server

Before this fix, we'd "render" error logs before sending them to
the server which meant that the "error" part was kept but the "msg" part
was dropped.

The user still sees the error but the message can be a bit chopped. This
fixes that by sending the "msg" field as is and adding an "error" field
to the payload as well that cloud handles separately.