fix(core): resolve templates in `source.path`
Before this fix, we weren't resolving template strings on the optional
`source.path` action config field. This caused errors when `git` was
called on a path that still had unresolved template strings.

This was fixed by preprocessing action configs (which resolves template
strings on them) before passing the paths to the VCS helper that
computes minimal roots.