feat: change default git scan mode to `repo` (#5399)
* feat: change default git scan mode to `repo`

The `repo` scan mode loads the entire repo files into a tree structure and applies glob matching from there.
This prevents a large amount of `git` subprocesses being spawned, and is much faster for large projects.
Especially on Windows it has a large performance benefit due to Windows being slow at spawning subprocesses.
We change this to be the default now, but it can always be overridden using the `GARDEN_GIT_SCAN_MODE=subtree` environment variable.

* chore: less noisy logging

* test: fix garden directory for version calculation tests