perf: do not compute log message until required (#5410)
* perf: enable logging pipeline to defer resolving log messages until needed

The idea is that sometimes we do heavy operations for silly logs, but we then discard just the log output.
Instead, we can pass a callback and if the log is never needed for display, it won't actually call the heavy operations.

* perf: convert heavy silly logs into deferred logs

* perf: add memoization to log callback

There might be multiple loggers resolving the message eventually.
Wrap it in `memoize` so that it only gets resolved once.

* fix: resolve messages in logs toString method

* test: fix log warning test