improvement(core): better action lifecycle logs + fix types
This commits adds more logging to the action lifecycle. That is logs
about getting status of a given action, the results, next steps, etc.

The logging is done in the task decorators that are also used for
emitting events.

The goal is obviously to make the logs more informative but also more
consistent.

Furthermore, this commit fixes the type for the `styles` map which were
way off. This means you can "break out" of the map and call all chalk
methods when chaining and it would be nice to narrow that down. But
we're going for loose and correct for now.