fix(dev): fix reload error when using templates (#5329)
* fix(dev): fix reload error when using templates

Before this fix, changing a Garden config during a `dev` command session
and running two subsequent commands would result in the first command
succeeding but the second failing with an error like:

"Build foo references template <some-template> which cannot be found.
Vailable templates: <None>"

This was because the config templates scanned by the first subcommand
after the config was changed weren't propagated to the parent Garden
instance, which was then cloned and passed to the second command (which
then had action configs, but no templates).

This was fixed by making the relevant instance variables on the Garden
class readonly (since reassignments in a child instance prevent newly
scanned configs from propagating to the parent via the shared data
structure) and emptying the maps before rescanning instead.

Also added a unit test case that checks this whole flow.

* chore(test): disable config clearing by default

The call to `this.clearConfig()` in `Garden#scanAndAddConfigs` had
started breaking certain test cases that manually set action configs
before calling `garden.getConfigGraph` (which would scan and add
configs, thus discarding the manually set configs).

While it's not great that we're adding more tweaks to the `TestGarden`
class, we're almost never interested in explicitly testing the post-edit
config reloading (outside of tests that test interactive / dev command
funcftionality).