docs: use cases headlines (#5214)
* docs: add Local Development with Remote Clusters

This adds a new user journey for users wanting to take advantage of Garden for its remote development features. This should be a how-to for the user at work who wants to get value for time quicky.

* docs: Local Development with Remote Clusters

This adds a new user journey for users wanting to take advantage of Garden for its remote development features. This should be a how-to for the user at work who wants to get value for time quicky.

* docs: add Portable Pipelines

* docs: deprecate use cases

There's fantastic material here but it's imposing and better disseminated throughout a user journey tied to Portable CI Pipelines.

* docs: update remote dev and CI use cases

* docs: add new jumpstart IDP doc

* fix: jumpstart idp title

* chore: docs gen, sidebar order

* chore: fix dead links

* chore: clarify IDP stack benefits

* chore: add hot reload and testing to local dev docs

Enhanced the local development documentation to include information
about using hot reload for smoother development loops and how to run
integration and end-to-end tests during the development process using
`garden test` or the Garden Web Dashboard.

* chore: update CI pipeline docs to address review

Refined the documentation on portable CI pipelines to address review.
Removed the prerequisites section to streamline the content and updated
the Next Steps to direct users to the quickstart guide. We can
re-introduce deleted sections when we build out docs journeys.

* docs: add caching mention to CI pipelines

* fix: dead links

* fix: dead links

---------

Co-authored-by: Tao Hansen <tao@garden.io>