improvement(logger): update some log lines after style changes (#5507)
The main issue was that the autocomplete suggestion was the same color
as the actual text which has now been fixed.

Beyond that, this replaces most calls to `accent` with `highlight` which
is the better choice for these lines.

It also changes the actual accent color to make it "pop" more. Before it
was the same as the primary color.

Finally I also used the chance to clean up some lines and ensure certain
log lines aren't printed when in dev mode or running `noProject`
commands.