docs: remove gif from quickstart guide + add sync flag
The gif is a bit distracting and pushes the actual content of the guide
below the initial view port which may discourage users from following along.

The gif is also not really relevant to the _quickstart_ guide, rather
it's a tour of the dashboard.

We now have that tour on the Dashboard docs page at:
https://docs.garden.io/using-garden/dashboard

The Quickstart example repo itself now also includes a gif that actually
demonstrates the steps of the guide:
https://github.com/garden-io/quickstart-example

So with the above in mind, I think we can remove this one.

Also, this commit adds the `--sync` flag to the deploy command in the
guide which is the intended behaviour.