refactor(mutagen): use faux SSH command to switch away from Mutagen fork
Garden currently uses a custom fork of Mutagen for its `kubectl exec`
transport. This is the last remaining blocker to get Garden back on the
official Mutagen releases.

At the moment, Mutagen doesn't support custom transports, but it does
allow a custom ssh executable to be specified, which presents an
opportunity for using a faux SSH command that calls down to `kubectl exec`
using parameters smuggled through the SSH hostname specification. In
this case, they are smuggled in a Base64-encoded JSON object (with some
slight encoding tweaks to pass through Mutagen's local path detection
heuristic).

The faux SSH implementation added in this commit will almost certainly
be useable (with minor tweaks) as a custom transport implementation once
support exists in Mutagen.

This commit is merely a DRAFT because it does not include a mechanism
for distribution of this faux SSH command or computation of its storage
location to specify via `MUTAGEN_SSH_PATH`.