fix(core): inherit Build action mode from dependant Deploy action (#5589)
* fix(core): reliably inherit build action mode from deploy action

Override the Build action mode with the Deploy action mode, if it depends on the build.

This enables using ${this.mode} in the build action as a proxy to the
Deploy action mode, simplifying configuration for many users.

In previous versions of Garden all actions went into `sync` Mode when
users ran the command `garden deploy --sync` without explicitly listing
build actions due to the fact that we defaulted to the `*` Minimatch
pattern that also matched build actions.

This commit ensures that if you depend on a Build from a Deploy action,
and that Deploy action is in `sync` mode, the Build action inherits the
mode.

If multiple Deploy actions depend on the same build, and have different
modes, the non-default mode wins. If multiple Deploy actions use
different non-default modes, we print a warning.

In the future, we can make this even more reliable by injecting
different invariants of build actions into the stack graph for each
Deploy action, if the build action is used by multiple Deploy actions
and they are in different modes.

* fix(deploy): only put Deploy actions into sync or local modes

When running `garden deploy --sync` without specifying certain actions
to be synced, previously all actions, even builds and runs went into
sync mode due to the minimatch pattern `*` matching all actions
regardless of Kind.

* test: add test for build mode inheritance from deploy actions

* docs: add a paragraph about mode inheritance in the actionModeSchema

* chore: undo accidental change in vote example

* chore: apply suggestion from code review

Co-authored-by: Thorarinn Sigurdsson <thorarinnsigurdsson@gmail.com>

---------

Co-authored-by: Thorarinn Sigurdsson <thorarinnsigurdsson@gmail.com>