fix(core): fix "repo" scan mode for remote actions
Before this fix, remote actions (i.e. actions with
`source.repository.url` set, not sources declared at the project level)
didn't work correctly when using the new `repo` git scanning mode.

Also set the default git scanning mode back to `repo`, since this bug
(which we've now fixed) was the reason we reverted to `subtree`.

The error arose because the base path wasn't being set to the locally
cloned path of the remote source's repo early enough in the resolution
process. This led to the repo scanning logic using the action's config
directory instead of the cloned source directory as the repo root
(thus ignoring the sources).

We've added a unit test that verifies that the base path is being set
correctly when the action graph is initialized.