fix(module-conversion): always include build deps
Before this fix, build dependencies weren't being transferred to the
generated actions in all cases.

This was a problem e.g. if using a `helm` module with a build dependency
on a `container` module, whose built image was referenced in the chart
(since the missing build dependency meant that the image wasn't
necessarily ready by the time the deploy started).

This was fixed by simply ensuring that any explicit build dependencies
(i.e. under `module.build.dependencies`) are included in the
dependencies of any generated runtime actions (Deploys, Tests and Runs).

This reflects the 0.12-era semantics, where every test, service and task
had a build dependency on the module's build step, which in turn
respected any build dependencies declared on the module (under
`build.dependencies`).