fix(module-conversion): always include build deps (#5671)
* fix(module-conversion): always include build deps

Before this fix, build dependencies weren't being transferred to the
generated actions in all cases.

This was a problem e.g. if using a `helm` module with a build dependency
on a `container` module, whose built image was referenced in the chart
(since the missing build dependency meant that the image wasn't
necessarily ready by the time the deploy started).

This was fixed by simply ensuring that any explicit build dependencies
(i.e. under `module.build.dependencies`) are included in the
dependencies of any generated runtime actions (Deploys, Tests and Runs).

This reflects the 0.12-era semantics, where every test, service and task
had a build dependency on the module's build step, which in turn
respected any build dependencies declared on the module (under
`build.dependencies`).

* chore: remove unnecessary null-checker

* chore: extract some local vars and fix typo

* fix(container): don't add build dep on PVCs

In the module conversion logic for the `persistentvolumeclaim` type, we
don't generate any Build actions (just a single Deploy).

This means we shouldn't declare a build dependency for it (since we'd be
declaring a dependency on an action that won't exist after the
conversion).

* chore(test): add logging env vars for testing

Sometimes, it's convenient to see the logger output while debugging test
suites. By default, we use the `quiet` logger (which doesn't log
anything).

Here, we introduce two new env vars for testing,
`GARDEN_TESTS_SHOW_LOGS` and `GAREDEN_TESTS_LOG_LEVEL`.

* test: update test assertions

---------

Co-authored-by: Vladimir Vagaytsev <vladimir.vagaitsev@gmail.com>