fix(module-conversion): more fixes to PVC type (#5681)
* fix(module-conversion): more fixes to PVC type

This is a follow-up to #5671 (see `47c24d5`). We also need to remove the
added dependencies on the build steps of `persistentvolumeclaim`
modules, since the converted actions there don't include a Build action.

Although this hasn't been reported yet, this would result in missing
dependency errors when executing Runs or Tests that come from
`container` modules which mount a `persistentvolumeclaim` module (using
the `volumes` config field for `container` module services).

* test: fix tests

---------

Co-authored-by: Vladimir Vagaytsev <vladimir.vagaitsev@gmail.com>