fix(git): fix file list caching bug for repo mode
This fixes an error that would come up e.g. when an action config in the
project root excluded a subdirectory containing another action config.

In that setup, the file list for the action in the subdirectory would be
empty because the file list had been cached using the exclude config
applied when computing the file list for the action config at the root
level (which included no files for the excluded subdirectory).

This was fixed by adding another level of caching to `GitRepoHandler`
that accounts for different includes/excludes and filter functions for
the same path.