docs: stop using build field in docs and examples (#5700)
When we released Bonsai we introduced the following short hand
for referencing container Build actions in container Deploy actions
via the `build` field:

```yaml
kind: Deploy
name: api
build: api
spec:
```

The longer version would be:

```yaml
kind: Deploy
name: api
dependencies: [build.api]
spec:
  image: ${actions.build.api.outputs.deploymentImageId}
  # ...
```

The short hand works for `container` Deploy actions but not for `helm`
or `kubernetes` Deploy actions. In fact using the `build` field there
sets the build context to the `.garden/build` dir of the corresponding
`build`.

This has caused a lot of confusion for users as they copy examples
containing the `build` field to `kubernetes` and `helm` Deploy actions.

A common symptom is that a Garden command will fail because it can't
find the files belonging to the Deploy action.

Furthermore, the `build` field has in general caused confusion and we
_may_ end up deprecating it.

In the meantime, this commit removes any unneeded/false usage of it in our
examples and docs.

The result is that `container` Deploy actions are slightly more
verbose but in turn we've removed a big footgun that users keep bumping
into as they copy/paste example config.

There are still legit use cases for the `build` field and I've left
those examples alone.