improvement(ext-tools): always print versions in `garden tools` output (#5718)
**What this PR does / why we need it**:
To improve UX, let's always print the concrete versions of the tools in
the `garden tools` output.

This PR also does refactoring in the tool definitions. It extracts tools
versions as named constants to make the version update process easier.
The PR also adds some missing tests to verify the binary hashes for the
tools.

**Which issue(s) this PR fixes**:

Fixes #

**Special notes for your reviewer**: