fix(cli): only overwrite terminal writer if using Ink (#5688)
<!--  Thanks for sending a pull request! Here are some tips for you:

1. If this is your first pull request, please read our contributor
guidelines in the
https://github.com/garden-io/garden/blob/main/CONTRIBUTING.md file.
2. Please label this pull request according to what type of issue you
are addressing (see "What type of PR is this?" below)
3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add `WIP:` at the beginning of the title or
use the GitHub Draft PR feature.
5. Please add at least two reviewers to the PR. Currently active
maintainers are: @edvald, @thsig, @eysi09, @shumailxyz, @stefreak,
@TimBeyer, @mkhq, and @vvagaytsev.
-->

**What this PR does / why we need it**:

Before this fix, we'd reset the terminal writer type after initialising
the command at the CLI level.

That's basically fine in and off itself except that it broke the test
logger and overwrote the settings we use there.

Instead of adding even more config options for the log level I opted for
simplifying the flow and only overwriting the type if using the 'Ink'
terminal write, since that's only use case for it anyway.

That implicitly fixes the issue with the test logger and slightly
simplifies the overall flow.

**Which issue(s) this PR fixes**:

Fixes #

**Special notes for your reviewer**: