fix(core): versioning fix for remote sources (#5735)
<!--  Thanks for sending a pull request! Here are some tips for you:

1. If this is your first pull request, please read our contributor
guidelines in the
https://github.com/garden-io/garden/blob/main/CONTRIBUTING.md file.
2. Please label this pull request according to what type of issue you
are addressing (see "What type of PR is this?" below)
3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add `WIP:` at the beginning of the title or
use the GitHub Draft PR feature.
5. Please add at least two reviewers to the PR. Currently active
maintainers are: @edvald, @thsig, @eysi09, @shumailxyz, @stefreak,
@TimBeyer, @mkhq, and @vvagaytsev.
-->

**What this PR does / why we need it**:

Before this fix, we were including the relative path to a file from
project root in the string to be hashed when calculating versions.

This was problematic when linking remote sources, since the chosen name
and path of the locally linked repo would be factored into the version
(resulting in unwanted cache misses).

This was fixed by using the relative path from the directory containing
the file's parent config. For example, an action's directory should be
able to be moved around or renamed without it affecting the version of
that action.

**Which issue(s) this PR fixes**:

Fixes #5501.