feat: added profile-project command
Added a simple utility command for profiling projects, the
`garden util profile-project` command.

This command logs the number of files included in each action and module
in the project (omitting actions that were converted from modules to
avoid duplication), their include/exclude configs (if any), and finally
the number of modules, actions and tracked files in the project.

This should be useful when helping users with large projects diagnose
slow init performance and how to best structure their includes,
excludes and `.gardenignore`.