feat: added profile-project command (#5780)
<!--  Thanks for sending a pull request! Here are some tips for you:

1. If this is your first pull request, please read our contributor
guidelines in the
https://github.com/garden-io/garden/blob/main/CONTRIBUTING.md file.
2. Please label this pull request according to what type of issue you
are addressing (see "What type of PR is this?" below)
3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add `WIP:` at the beginning of the title or
use the GitHub Draft PR feature.
5. Please add at least two reviewers to the PR. Currently active
maintainers are: @edvald, @thsig, @eysi09, @stefreak, @TimBeyer, and
@vvagaytsev.
-->

**What this PR does / why we need it**:

Added a simple utility command for profiling projects, the
`profile-project` command.

This command logs the number of files included in each action and module
in the project (omitting actions that were converted from modules to
avoid duplication), their include/exclude configs (if any), and finally
the number of modules, actions and tracked files in the project.

This should be useful when helping users with large projects diagnose
slow init performance and how to best structure their includes, excludes
and `.gardenignore`.