feat(modules): allow opting out of build staging
Here, we make the `local` config field (previously only available to
`exec` modules) available for all module types.

When used, it maps to the `buildAtSource` field of any generated Build
actions in the module conversion process.

This is done because some users/projects need to reference symlinked
directories in their builds, where the simplest solution appears to be
to simply opt out of build staging for those modules.