improvement(k8s): better pod runner error handling
Before this fix, we weren't properly propagating pod runner errors when
performing status checks for in-cluster builds.

We also make some tweaks to the logic that checks if the runner pod was
killed. Instead of using the local time at the start of the pod run as
the reference/cutoff, we now use the latest event timestamp before
exec-ing (which should make the logic robust against timezone
differences).

Co-authored with @stefreak