fix(communityEdition): fix unprocessable entity error (#5931)
Fix an issue where garden failed with the error "unprocessable entity"
some times when user has more than 24 projects.

Previously core listed all projects and if it didn't find project with
given name in the first list of results, it would attempt creating the
project. Creating the project fails then with "unprocessable entity"
because a project with the given name already exists.