refactor(git): git handler class hierarchy (#5967)
* chore: declare internally used methods as private

* chore: rename `GitHandler` to `GitSubTreeHandler`

To align naming with another components and functions,
and to make it more descriptive.

* refactor: extract class `AbstractGitHandler`

It has the only abstract method `getFiles(...)`
that would be implemented by each Git repo scanning mode.

* refactor: move `GitSubTreeHandler` impl to a separate file

* refactor:  replace inheritance with composition

Now `GitRepoHandler` extends `AbstractGitHandler` instead of `GitSubTreeHandler`,
and contains the latter as a class member
to be able to delegate to its `getFiles()` method if necessary.

* chore: fix imports

* refactor: declare `VcsHandler.name` as readonly

* chore: linting

* chore: fix imports

* refactor: remove unnecessary hack to avoid infinite recursion