perf(cli): avoid unnecessary module resolution when filtering by name (#6002)
* perf(cli): avoid unnecessary module resolution when filtering by name

This applies to most common usages of the `build`, `deploy`, `run` and `test`
commands when one or more names are specified, as well as in the
`get modules` and `get actions` commands.

For now this is enabled specifically with by setting the
`GARDEN_ENABLE_PARTIAL_RESOLUTION=true` env variable.

* perf: extend partial resolution to action graph resolution

* chore(test): fix compile-time errors

Post-rebase fix

* chore: unify log format

Print class name as a prefix in all debug/silly messages.

* refactor: simplify for-loop

* fix: use correct task type in log messages

* chore: more detailed error message on available actions

* fix: throw if no actions found by pattern matching

* Revert "fix: throw if no actions found by pattern matching"

This reverts commit 4b86e055e971326c6957a5358bb00c1e068bf149.

* fix: ugly hack to avoid changes in behaviour of build and deploy commands

---------

Co-authored-by: Vladimir Vagaytsev <vladimir.vagaitsev@gmail.com>