fix: `garden publish` command to respect `publishId` (#6052)
* refactor: rename local vars for more clarity

* chore: log image tags while publishing

* chore: (tidying) re-arrange the code

To see more similarities between 2 handlers

* fix: read image tag from existing parsed image

The parsed image comes from a helper function.
That helpers function uses `splitLast` instead of `splitFirst`
with `:`-delimiter to get the tag name. That looks correct.

* chore: rename some local vars for more clarity

* chore: optimized imports

* refactor: inline local var

* refactor: extract helper function to reduce code duplication

The only diff between `getContainerBuildActionOutputs` and `k8sGetContainerBuildActionOutputs`
was in the deploymentImageId calculation that depended
on the presence of deployment registry config.

* chore: remove outdated comment

* refactor: rename local vars for more clarity

* refactor: rename functions args for more clarity

* chore: add clarifying comment

* refactor: unwrap else-conditions

* chore: always tag container images

Align the behaviour with the `k8sPublishContainerBuild`
which does not compare local and remote images.

The repeated tagging is idemponent and should not be too costly.

* refactor: inline function

It was used only in one place.

* chore: amend var names, logs and comments for better clarity

* chore: log Docker tags while local build

* refactor: declare semantically different image ids separately

Each in its own variable.

* fix: use correct Docker image ID in `publish` command

* fix: use local image name for deployment registries

Publish ID is used only by `publish` command,
and it should not be used for internal deployment registries
and while Build-actions resolution.

* improvement: add image details to outputs of `publish` command

* chore: delete outdated comment