improvement(k8s): allow volume mounts in runners
If a `kubernetes-pod` Run or Test includes `volumeMounts` on the first
container in its `podSpec`, we now include this in the runner pod.

Previously, we omitted the `volumeMounts` field from the first container
in the (optional) user-supplied `podSpec` field. This is because the
relevant code assumed `container` Runs/Tests—it's now appropriate for
`container` and `kubernetes-pod` Runs/Tests.