fix(ext-tools): remove default `cwd` for external tools
All Garden bundled tools' paths are known to be absolute.
There is no reason to set `cwd` for running any bundled tool,
because it does not have any effect.

But, it can cause some troubles with particular tools.
The latest kubectl, that uses Go 1.19+, can fail,
because of the change made in Go 1.19,
see https://pkg.go.dev/os/exec#hdr-Executables_in_the_current_directory

So, it is safer to remove the unnecessary default `cwd` definition.