test(git): fix test-cases for post-Git-scan FS modifications (#6198)
* test(git): clean test setup to uncover the issue

This test fails because the results are cached.
To imitate Garden CLI rerun we need to re-create the Git handler in the tests,
or to invalidate its cache.

* test(git): move non-mocked `getTreeVersion` tests to `git.ts`

This will ensure that these tests are run in both scanning modes.

* test(git): add one more test config

To test modification to already scanned files with different project structure.
The project config structure can affect the repo scanning roots and the cache hits.

* test(git): reset VCS caches when necessary